back-end/shopify_apps/shoppingcart/serializers.py
from rest_framework import serializers
from .models import CartItem, ShoppingCart

class CartItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = CartItem
        fields = ['item', 'date']

class ShoppingCartSerializer(serializers.ModelSerializer):
    cart_items = CartItemSerializer(many=True, read_only=True)

    class Meta:
        model = ShoppingCart
        fields = ['cart_items']