back-end/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.0
Django==4.2.6
django-appconf==1.0.6
django-environ==0.11.2
django-filter==23.5
django-measurement==3.2.4
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
Markdown==3.5.1
MarkupSafe==2.1.3
measurement==3.2.2
model-bakery==1.17.0
mpmath==1.3.0
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-environ==0.4.54
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.16.2
simplejson==3.19.2
sqlparse==0.4.4
sympy==1.12
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
